+++ crates/puniyu_config/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "puniyu_config"
+version = "0.1.0"
+authors.workspace = true
+license.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+notify = "8.2.0"
+flume = "0.11.1"
+
+puniyu_logger.workspace = true
+serde.workspace = true
+
+puniyu_common = { version = "0.1.0", path = "../puniyu_common" }
+
